Reorganized file structure, updated node packages, removed firebase files
