--- .firebaserc
@@ -1,5 +1,0 @@
-{
-  "projects": {
-    "default": "personal-website-19d67"
-  }
-}
