--- boxRoulette/boxRoulette.js
@@ -1,5 +1,0 @@
-import { createApp } from "vue";
-import "../src/features/boxRoulette/style.css";
-import App from "../src/features/boxRoulette/App.vue";
-
-createApp(App).mount("#app");
