--- boxRoulette/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="./j-square.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Box Roulette</title>
-  </head>
-  <body>
-    <div id="app"></div>
-    <script type="module" src="boxRoulette.js"></script>
-  </body>
-</html>
