--- homepage.js
@@ -1,5 +1,0 @@
-import { createApp } from "vue";
-import "./src/base.css";
-import App from "./src/features/homepage/App.vue";
-
-createApp(App).mount("#app");
