+++ index.html
@@ -27,6 +27,6 @@
   <!-- Google tag (gtag.js) END -->
   <body>
     <div id="app"></div>
-    <script type="module" src="homepage.js"></script>
+    <script type="module" src="/src/main.js"></script>
   </body>
 </html>
