+++ package.json
@@ -10,14 +10,15 @@
   },
   "dependencies": {
     "lucide-vue-next": "^0.544.0",
-    "vue": "^3.5.18"
+    "vue": "^3.5.18",
+    "vue-router": "^4.6.4"
   },
   "devDependencies": {
     "@eslint/js": "^9.33.0",
-    "@vitejs/plugin-vue": "^6.0.1",
+    "@vitejs/plugin-vue": "^6.0.7",
     "eslint": "^9.33.0",
     "eslint-plugin-vue": "^10.4.0",
     "globals": "^16.3.0",
-    "vite": "^7.1.2"
+    "vite": "^8.0.16"
   }
 }
