+++ src/features/boxRoulette/data.js
@@ -1,0 +1,4 @@
+export const project = {
+  title: 'Box Roulette',
+  summary: 'A gambling-style box-picking game built with Vue.js.',
+}
