+++ src/features/boxRoulette/index.js
@@ -1,0 +1,2 @@
+export { default as BoxRouletteView } from './views/BoxRouletteView.vue'
+export { project } from './data.js'
