--- src/features/boxRoulette/style.css
@@ -1,89 +1,0 @@
-:root {
-  font-family: Poppins, system-ui, Avenir, Helvetica, Arial, sans-serif;
-
-  /* color-scheme: light dark; */
-  color: white;
-  background-color: black;
-}
-/* CSS Reset By Josh Comeau https://www.joshwcomeau.com/css/custom-css-reset/ */
-/*
-  1. Use a more-intuitive box-sizing model.
-*/
-*,
-*::before,
-*::after {
-  box-sizing: border-box;
-}
-/*
-  2. Remove default margin
-*/
-* {
-  margin: 0;
-}
-/*
-  Typographic tweaks!
-  3. Add accessible line-height
-  4. Improve text rendering
-*/
-body {
-  line-height: 1.5;
-  -webkit-font-smoothing: antialiased;
-}
-/*
-  5. Improve media defaults
-*/
-img,
-picture,
-video,
-canvas,
-svg {
-  display: block;
-  max-width: 100%;
-}
-/*
-  6. Remove built-in form typography styles
-*/
-input,
-button,
-textarea,
-select {
-  font: inherit;
-}
-/*
-  7. Avoid text overflows
-*/
-p,
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  overflow-wrap: break-word;
-}
-/*
-  8. Create a root stacking context
-*/
-#root,
-#__next {
-  isolation: isolate;
-}
-/* Reset End */
-
-button {
-  color: white;
-  transition: all 0.2s ease-in-out;
-  background-color: #ffffff00;
-  border: none;
-}
-button:hover {
-  cursor: pointer;
-}
-body {
-  margin: 0;
-  display: flex;
-  flex-direction: column;
-  justify-content: start;
-  min-width: 250px;
-  min-height: 100vh;
-}
