--- src/features/homepage/App.vue
@@ -1,15 +1,0 @@
-<script setup>
-import NavBar from "./components/NavBar.vue";
-import FooterBar from "./components/FooterBar.vue";
-import HeroSection from "./components/HeroSection.vue";
-import ProjectSection from "./components/ProjectSection.vue";
-</script>
-
-<template>
-  <NavBar />
-  <HeroSection />
-  <ProjectSection />
-  <FooterBar />
-</template>
-
-<style scoped></style>
