--- src/features/homepage/scripts/socialLinks.js
@@ -1,26 +1,0 @@
-import { Github, Instagram, Linkedin, Mail } from "lucide-vue-next";
-
-const links = [
-  {
-    alt: "Github",
-    url: "https://github.com/McJet",
-    svg: Github,
-  },
-  {
-    alt: "LinkedIn",
-    url: "https://www.linkedin.com/in/jetherysalina",
-    svg: Linkedin,
-  },
-  {
-    alt: "Email",
-    url: "mailto:jethercy@gmail.com",
-    svg: Mail,
-  },
-  {
-    alt: "Instagram",
-    url: "https://www.instagram.com/jether_y/",
-    svg: Instagram,
-  },
-];
-
-export default links;
