+++ src/features/landing/data/socialLinks.js
@@ -1,0 +1,8 @@
+import { Github, Instagram, Linkedin, Mail } from "lucide-vue-next";
+
+export const socialLinks = [
+  { alt: "Github", url: "https://github.com/McJet", svg: Github },
+  { alt: "LinkedIn", url: "https://www.linkedin.com/in/jetherysalina", svg: Linkedin },
+  { alt: "Email", url: "mailto:jethercy@gmail.com", svg: Mail },
+  { alt: "Instagram", url: "https://www.instagram.com/jether_y/", svg: Instagram },
+]
