+++ src/features/landing/styles/variables.css
@@ -1,0 +1,41 @@
+:root {
+  color-scheme: light dark;
+  --font-family-header: "Poppins", sans-serif;
+  --font-family: "Hind", sans-serif;
+  --font-size-base: 16px;
+
+  --white-hex: #eeeeee;
+  --white-rgb: 238, 238, 238;
+  --light-gray-hex: #d5d5d5;
+  --gray-hex: #bfc0c0;
+  --gray-rgb: 191, 192, 192;
+  --blue-hex: #4f5d75;
+  --blue-rgb: 79, 93, 117;
+  --dark-blue-hex: #2d3142;
+  --dark-blue-rgb: 45, 49, 66;
+  --black-hex: #07071b;
+  --black-rgb: 7, 7, 27;
+  --orange-hex: #ef8354;
+  --orange-rgb: 239, 131, 84;
+
+  --color-bg: light-dark(var(--white-hex), var(--dark-blue-hex));
+  --color-bg-secondary: light-dark(#e0e0e0, #363c55);
+  --color-text: light-dark(var(--dark-blue-hex), var(--white-hex));
+  --color-text-secondary: light-dark(var(--blue-hex), var(--light-gray-hex));
+  --color-secondary: var(--orange-hex);
+  --color-border: light-dark(
+    rgba(var(--gray-rgb), 0.75),
+    var(--color-text-secondary)
+  );
+
+  --transition: 0.2s color, 0.2s background-color, 0.2s border-color;
+
+  font-family: var(--font-family);
+  font-size: var(--font-size-base);
+  letter-spacing: 0.01rem;
+  line-height: 150%;
+  text-rendering: optimizeLegibility;
+
+  background-color: var(--color-bg);
+  color: var(--color-text);
+}
