--- src/features/test-page/App.vue
@@ -1,9 +1,0 @@
-<script setup>
-import FontTest from "./FontTest.vue";
-</script>
-
-<template>
-  <FontTest />
-</template>
-
-<style scoped></style>
