+++ src/features/test-page/data.js
@@ -1,0 +1,4 @@
+export const project = {
+  title: 'Test Page',
+  summary: 'A playground for testing UI components and typography.',
+}
