+++ src/features/test-page/index.js
@@ -1,0 +1,2 @@
+export { default as TestPageView } from './views/TestPageView.vue'
+export { project } from './data.js'
