+++ src/features/test-page/views/TestPageView.vue
@@ -1,0 +1,8 @@
+<script setup>
+import '@/features/landing/styles/main.css'
+import FontTest from '../components/FontTest.vue'
+</script>
+
+<template>
+  <FontTest />
+</template>
