+++ src/main.js
@@ -1,0 +1,6 @@
+import { createApp } from 'vue'
+import App from './App.vue'
+import router from './router'
+import './shared/styles/reset.css'
+
+createApp(App).use(router).mount('#app')
