--- test-page/index.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="./j-square.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Test Page</title>
-    <link
-      href="https://api.fontshare.com/v2/css?f[]=hind@400&f[]=poppins@600,700&display=swap"
-      rel="stylesheet"
-    />
-  </head>
-  <!-- Google tag (gtag.js) -->
-  <!-- <script
-    async
-    src="https://www.googletagmanager.com/gtag/js?id=G-ZLP9J3JHFG"
-  ></script>
-  <script>
-    window.dataLayer = window.dataLayer || [];
-    function gtag() {
-      dataLayer.push(arguments);
-    }
-    gtag("js", new Date());
-
-    gtag("config", "G-ZLP9J3JHFG");
-  </script> -->
-  <!-- Google tag (gtag.js) END -->
-  <body>
-    <div id="app"></div>
-    <script type="module" src="testPage.js"></script>
-  </body>
-</html>
